update python files

+++ clean.py
@@ -1,9 +1,11 @@
 import os
 
-def clean():
+def clean(exception="conf.ndjson"):
     trace_files = [f for f in os.listdir(".") if f.endswith('.ndjson')]
     print(f"Cleanup: {trace_files}")
     for trace_file in trace_files:
+        if trace_file == exception:
+            continue
         os.remove(trace_file)
 
 if __name__ == "__main__":
